:art: refactor: URL 수정 및 authController 분리
URL 수정
- url에 버전 명시
- 중복검사 url 직관적으로 변경
- RequestParam을 사용하려 했으나 중복되는 엔드포인트와 swagger문제가 발생하여 PathVariable로 유지

AuthController
- 로그인, 회원가입 기능 분리
- 추가적으로 로그아웃 기능 개발
